rename and move filter to table-filters
